Refactor: Move the GND lightmap export to RagnarokTools
It's a grab bag of random utilities for now, but at least it's unit tested.
